Create PrductPriceSimulationWrapper
