+++ react/__mocks__/vtex.store-resources/queries/itemsWithSimulation.gql
@@ -1,5 +1,5 @@
 query itemsWithSimulation($items: [ItemInput]) {
-  itemsWithSimulation(items: $items) {
+  itemsWithSimulation(items: $items) @context(provider: "vtex.search-graphql") {
     itemId
     sellers {
       commertialOffer {
